Hi! I just discovered this package and it's amazing! I just create a SCGLEkernel.jl to include the kernel associated with the Self-Consistent Generalized Langevin Equation Theory.
Thanks to the examples in the docs, my extension was very straightforward. The F function of the integrodifferential problem is composed by a concatenated array for the self-component and an array for the collective-component of the Intermediate Scattering Function. So the initial F(0) = (1,S)^T where 1 is a unit array of the length of the static structure factor array S. I include an usage example at the bottom of the file.